Refactor: Extract dateFormat function into individual file
